rename hierarchy imports in worker.*
