+++ core/src/main/java/tachyon/worker/nio/NIODataServer.java
@@ -37,7 +37,7 @@ import tachyon.Constants;
 import tachyon.conf.TachyonConf;
 import tachyon.worker.BlocksLocker;
 import tachyon.worker.DataServer;
-import tachyon.worker.hierarchy.StorageDir;
+import tachyon.worker.tiered.StorageDir;
 
 /**
  * The Server to serve data file read requests from remote machines. The current implementation is
